renterutil: Remove UploadDir
This means that we lose file-packing, one of the big advantages of
us that I'd been touting. Not to worry; it will return when sectors
are properly buffered in the FileSystem. We'll need to do this anyway
to get a performant writeable FUSE, so the old functionality should
"fall out" as a natural consequence of a more general algorithm.

Actually writing such an algorithm, though, is rather daunting. I've
been putting it off for a while. :|