renterutil: MVP writable filesystem
This is horribly messy, but it's passing basic tests! I don't yet
have a test that checks whether sectors are being aggregated
properly (as in UploadDir), but sticking Printlns in manually seems
to confirm that they are indeed being aggregated. Which means we
should be back to full parity with the old non-filesystem code.

More importantly, we now have all the necessary pieces to implement
arbitrary writes! Once I wrap this in a thin FUSE layer, it should
be possible to run siad on a Sia FUSE mount (although I expect
performance will be terrible).

Caching seems to be working too: if you write a few bytes, then seek
backwards and try to read them, you'll read directly from memory
instead of having to download anything. Similarly, you can overwrite
previous uncommitted writes instantly.

There's a lot of room for improvement (some functions are really
massive and definitely need a helper or two, and I'm allocating
memory pretty carelessly and redoing work) but overall I'm very
excited about this. :)