shard: Allow host lookup by prefix
I decided to keep the existing pubkey->announcement map around, so
we'll get O(1) performance for full pubkey lookups (which are the
common case anyway) at the cost of doubling the amount of storage.
There are <16000 unique pubkeys at this point, so that's about 1MB
extra storage. Not exactly worth obsessing over!