user: Implement writable FUSE bindings
There's still a lot of functionality missing (e.g. Chmod), but this
is quite a milestone! Testing revealed that there are a lot of ugly
edge cases in filesystems, mainly around having the same file opened
with two different sets of flags. Nothing insurmountable though.

The only real release-blocker is the read performance: PseudoFiles
don't do any buffering or caching for reads (unlike writes), and it
seems that most applications prefer many small reads to a few large
reads. So I'll need to add some form of buffering, much like I did
for the HTTP server, in order to attain acceptable speeds.